refactor: move report-webhook Secret into Config controller
Move reconcileReportWebhookSecret and all report rendering logic from
ReportProcessorReconciler into ConfigReconciler. This aligns with the
existing pattern where the Config controller owns all generated
Secrets and ConfigMaps (like the ENC Secret).

The ReportProcessor controller now only logs reconcile events. Status
updates on ReportProcessor objects are handled by the Config
controller during Secret reconciliation.

Closes #55
